requirements.txt
aiogram==3.1.1
redis
asyncpg
gspread
google-auth
Pillow
python-dotenv
apscheduler
aiohttp
google-generativeai==0.5.0
